+++ modules/workflow-engine/src/main/java/io/mateu/workflow/infra/out/persistence/StepExecutionEntityRepository.java
@@ -7,4 +7,5 @@ import java.util.List;
 public interface StepExecutionEntityRepository extends JpaRepository<StepExecutionEntity, String> {
     List<StepExecutionEntity> findAllByProcessIdOrderByOrder(String processId);
     List<StepExecutionEntity> findAllByStatusIn(List<String> statuses);
+    List<StepExecutionEntity> findAllByProcessIdAndStatusIn(String processId, List<String> statuses);
 }